rptest: move load_tier_profiles() to redpanda_cloud.py
